backend/requirements.txt
fastapi
uvicorn

# for databases
sqlalchemy
psycopg2

# for env var
python-dotenv

# for email validation
pydantic[email]

#hashing
passlib[bcrypt]

#for testing
pytest
requests
pytest-cov==2.11.1

#for authentication
python-jose

#for form data
python-multipart

#for templates
jinja2

#for static files
aiofiles